Fix PR #95

+++ database/migrations/2020_08_01_000004_create_plan_subscription_usage_table.php
@@ -18,7 +18,7 @@ class CreatePlanSubscriptionUsageTable extends Migration
             $table->increments('id');
             $table->integer('subscription_id')->unsigned();
             $table->integer('feature_id')->unsigned();
-            $table->smallInteger('used')->unsigned();
+            $table->integer('used')->unsigned();
             $table->dateTime('valid_until')->nullable();
             $table->string('timezone')->nullable();
             $table->timestamps();
